chore(e2e): move e2e.prod to working with app-scripts
move shared functions into util, add support for debug flag when
running e2e.prod / demos.prod which gets passed to app-scripts
