+++ scripts/e2e/e2e.template.prod.html
@@ -43,7 +43,7 @@
       document.body.classList.remove('rtl');
     }
   </script>
-  <script src="../../../../polyfills/polyfills.ng.js"></script>
+  <script src="../../../../polyfills/polyfills.js"></script>
   <script src="./build/main.js"></script>
 </body>
 </html>
